+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Created the configuration dialog to allow the user to change the application settings.
 - Added splash screen and multi-language support to the desktop application.
 - Created the about dialog to show the application information on the desktop.
+- Created the new project dialog to allow the user to create a new project.
 
 ### Fix